New Development: Yes             JIRA Issue: CAS-1795       Ready to Release: No      Interface Changes: Yes What Interface Changed: MS2ASDM.h          Test Programs: test_task_exportasdm.py   Put in Release Notes: No              Module(s): alma            Description: Fixed several small but important details 			in the field and source tables and in 			the writing of the binary data. 			This is the first version which creates 			a complete and re-importable ASDM.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@10288 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/alma/implement/MS2ASDM/MS2ASDM.h
@@ -140,7 +140,7 @@ namespace casa { //# NAMESPACE CASA - BEGIN
     return (Double) atime.get() / (Double)ArrayTime::unitsInASecond; }
 
   asdm::Interval ASDMInterval( const Double seconds ){ 
-    return asdm::Interval((long) (floor(seconds*ArrayTime::unitsInASecond))); }
+    return asdm::Interval((long long) (floor(seconds*ArrayTime::unitsInASecond))); }
 
   // convert MS style time interval to ASDM ArrayTimeInterval
   asdm::ArrayTimeInterval ASDMTimeInterval( const Quantity midpoint, const Quantity interval);
